requirements-lock.txt
absl-py==2.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
black==24.8.0
Box2D==2.3.10
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
claude-code==0.0.1
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
distlib==0.4.0
etils==1.13.0
Farama-Notifications==0.0.4
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
glfw==2.9.0
gymnasium==1.2.0
hydra-core==1.3.2
identify==2.6.13
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
iniconfig==2.1.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
moviepy==2.2.1
mpmath==1.3.0
mujoco==3.3.5
mypy==1.11.2
mypy_extensions==1.1.0
networkx==3.5
nodeenv==1.9.1
npm==0.1.1
numpy==2.3.2
omegaconf==2.3.0
optional-django==0.1.0
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.4.0
pluggy==1.6.0
pre-commit==3.8.0
proglog==0.1.12
protobuf==6.32.0
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.6.1
PyOpenGL==3.1.10
pyparsing==3.2.3
pytest==8.3.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
ruff==0.6.4
sentry-sdk==2.36.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
stable_baselines3==2.7.0
swig==4.3.1.post0
sympy==1.14.0
torch==2.8.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.2.3
virtualenv==20.34.0
wandb==0.21.3
zipp==3.23.0
